refactor: migrate production ipc-contract imports to ipc-protocol
