+++ assistant/src/notifications/adapters/macos.ts
@@ -13,7 +13,7 @@
  * does not match their own identity.
  */
 
-import type { ServerMessage } from "../../daemon/ipc-contract.js";
+import type { ServerMessage } from "../../daemon/ipc-protocol.js";
 import { getLogger } from "../../util/logger.js";
 import type {
   ChannelAdapter,
